requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4
Faker==15.1.1
fancycompleter==0.9.1
freezegun==1.2.2
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.8.6
py==1.11.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrepl==0.9.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.3
testfixtures==6.18.5
tomli==2.0.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uWSGI==2.0.20
wmctrl==0.4
zipp==3.9.0
